refactor: Move lifecycle tests to colocated test file
Move the 'Lando App Lifecycle Commands Tests' suite from extension.test.ts
to a new lifecycle.test.ts file, following the project's pattern of
organizing feature-specific integration tests in their own files.

This aligns with the code refactoring that moved lifecycle command
implementations to src/commands/lifecycle.ts.
